+++ R/class-workbook.R
@@ -801,7 +801,8 @@ wbWorkbook <- R6::R6Class(
     #' @param withFilter withFilter
     #' @param name name
     #' @param sep sep
-    #' @param removeCellStyle removeCellStyle
+    #' @param removeCellStyle if writing into existing cells, should the cell style be removed?
+    #' @param na.strings na.strings
     #' @param return The `wbWorkbook` object
     add_data = function(
         sheet           = current_sheet(),
@@ -816,8 +817,12 @@ wbWorkbook <- R6::R6Class(
         withFilter      = FALSE,
         name            = NULL,
         sep             = ", ",
-        removeCellStyle = TRUE
+        removeCellStyle = FALSE,
+        na.strings
       ) {
+
+      if (missing(na.strings)) na.strings <- substitute()
+
       write_data(
         wb              = self,
         sheet           = sheet,
@@ -832,7 +837,8 @@ wbWorkbook <- R6::R6Class(
         withFilter      = withFilter,
         name            = name,
         sep             = sep,
-        removeCellStyle = removeCellStyle
+        removeCellStyle = removeCellStyle,
+        na.strings      = na.strings
       )
       self
     },
@@ -854,6 +860,7 @@ wbWorkbook <- R6::R6Class(
     #' @param lastColumn lastColumn
     #' @param bandedRows bandedRows
     #' @param bandedCols bandedCols
+    #' @param na.strings na.strings
     #' @returns The `wbWorkbook` object
     add_data_table = function(
         sheet       = current_sheet(),
@@ -871,8 +878,12 @@ wbWorkbook <- R6::R6Class(
         firstColumn = FALSE,
         lastColumn  = FALSE,
         bandedRows  = TRUE,
-        bandedCols  = FALSE
+        bandedCols  = FALSE,
+        na.strings
     ) {
+
+      if (missing(na.strings)) na.strings <- substitute()
+
       write_datatable(
         wb          = self,
         sheet       = sheet,
@@ -890,7 +901,8 @@ wbWorkbook <- R6::R6Class(
         firstColumn = firstColumn,
         lastColumn  = lastColumn,
         bandedRows  = bandedRows,
-        bandedCols  = bandedCols
+        bandedCols  = bandedCols,
+        na.strings  = na.strings
       )
       self
     },