+++ R/write.R
@@ -9,6 +9,7 @@
 #' @param data_class optional data class object
 #' @param colNames if TRUE colNames are passed down
 #' @param removeCellStyle keep the cell style?
+#' @param na.strings optional na.strings argument. if missing #N/A is used. If NULL no cell value is written, if character or numeric this is written (even if NA is part of numeric data)
 #'
 #' @examples
 #'    xlsxFile <- system.file("extdata", "update_test.xlsx", package = "openxlsx2")
@@ -32,7 +33,8 @@
 #'
 #' @export
 update_cell <- function(x, wb, sheet, cell, data_class,
-                        colNames = FALSE, removeCellStyle = FALSE) {
+                        colNames = FALSE, removeCellStyle = FALSE,
+                        na.strings) {
 
   dims <- dims_to_rowcol(cell)
   cols <- dims[[1]]
@@ -94,8 +96,7 @@ update_cell <- function(x, wb, sheet, cell, data_class,
       total_cols <- int2col(sort(col2int(total_cols)))
 
       # create candidate
-      cc_row_new <- data.frame(matrix("", nrow = length(total_cols), ncol = 3))
-      names(cc_row_new) <- names(cc)[1:3]
+      cc_row_new <- create_char_dataframe(names(cc)[1:3], length(total_cols))
       cc_row_new$row_r <- row
       cc_row_new$c_r <- total_cols
       cc_row_new$r <- stringi::stri_join(cc_row_new$c_r, cc_row_new$row_r)
@@ -158,8 +159,13 @@ update_cell <- function(x, wb, sheet, cell, data_class,
 
         # for now convert all R-characters to inlineStr (e.g. names() of a data frame)
         if ((data_class[m] == openxlsx2_celltype[["character"]]) || ((colNames == TRUE) && (n == 1))) {
-          cc[sel, "c_t"] <- "inlineStr"
-          cc[sel, "is"]   <- paste0("<is><t>", as.character(value), "</t></is>")
+          if (is.na(value)) {
+            cc[sel, "v"] <- "#N/A"
+            cc[sel, "c_t"] <- "e"
+          } else {
+            cc[sel, "c_t"] <- "inlineStr"
+            cc[sel, "is"]   <- paste0("<is><t>", as.character(value), "</t></is>")
+          }
         } else if (data_class[m] == openxlsx2_celltype[["formula"]]) {
           cc[sel, "c_t"] <- "str"
           cc[sel, "f"] <- as.character(value)
@@ -167,18 +173,42 @@ update_cell <- function(x, wb, sheet, cell, data_class,
           cc[sel, "f"] <- as.character(value)
           cc[sel, "f_t"] <- "array"
           cc[sel, "f_ref"] <- cell
-        }else if (data_class[m] == openxlsx2_celltype[["hyperlink"]]) {
+        } else if (data_class[m] == openxlsx2_celltype[["hyperlink"]]) {
           cc[sel, "f"] <- as.character(value)
           # FIXME assign the hyperlinkstyle if no style found. This might not be
           # desired. We should provide an option to prevent this.
-          if (cc[sel, "c_s"] == "")
+          if (cc[sel, "c_s"] == "" || is.na(cc[sel, "c_s"]))
             cc[sel, "c_s"] <- wb$styles_mgr$get_xf_id("hyperlinkstyle")
         } else {
-          cc[sel, "v"]   <- as.character(value)
+          if (is.na(value)) {
+            if (missing(na.strings)) {
+              cc[sel, "v"] <- "#N/A"
+              cc[sel, "c_t"] <- "e"
+            } else {
+              if (is.null(na.strings)) {
+                next # do not add any value: <c/>
+              } else {
+               cc[sel, "c_t"] <- "inlineStr"
+               cc[sel, "is"] <- txt_to_is(as.character(na.strings),
+                                          no_escapes = TRUE, raw = TRUE)
+              }
+            }
+          } else if (value == "NaN") {
+            cc[sel, "v"] <- "#VALUE!"
+            cc[sel, "c_t"] <- "e"
+          } else if (value == "-Inf" || value == "Inf") {
+            cc[sel, "v"] <- "#NUM!"
+            cc[sel, "c_t"] <- "e"
+          } else {
+            cc[sel, "v"]   <- as.character(value)
+          }
         }
 
       }
     }
+
+    # avoid missings in cc
+    cc[is.na(cc)] <- ""
 
   }
 
@@ -217,6 +247,7 @@ nmfmt_df <- function(x) {
 #' @param startRow row to place it
 #' @param startCol col to place it
 #' @param removeCellStyle keep the cell style?
+#' @param na.strings optional na.strings argument. if missing #N/A is used. If NULL no cell value is written, if character or numeric this is written (even if NA is part of numeric data)
 #' @details
 #' The string `"_openxlsx_NA"` is reserved for `openxlsx2`. If the data frame
 #' contains this string, the output will be broken.
@@ -241,8 +272,10 @@ nmfmt_df <- function(x) {
 write_data2 <-function(wb, sheet, data, name = NULL,
                        colNames = TRUE, rowNames = FALSE,
                        startRow = 1, startCol = 1,
-                       removeCellStyle = FALSE) {
+                       removeCellStyle = FALSE,
+                       na.strings) {
 
+  if (missing(na.strings)) na.strings <- substitute()
 
   is_data_frame <- FALSE
   #### prepare the correct data formats for openxml
@@ -365,16 +398,11 @@ write_data2 <-function(wb, sheet, data, name = NULL,
 
     # original cc data frame
     # TODO should be empty_sheet_data(n = nrow(data) * ncol(data))
-    nams <- c("row_r", "c_r", "c_s", "c_t", "c_cm",
-              "c_ph", "c_vm", "v", "f", "f_t",
-              "f_ref", "f_ca", "f_si", "is", "typ",
-              "r")
-    cc <- as.data.frame(
-      matrix(data = "",
-             nrow = nrow(data) * ncol(data),
-             ncol = length(nams))
-    )
-    names(cc) <- nams
+    nams <- c("r", "row_r", "c_r", "c_s", "c_t", "c_cm", "c_ph", "c_vm",
+              "v", "f", "f_t", "f_ref", "f_ca", "f_si", "is", "typ")
+
+    cc <- create_char_dataframe(colnames = nams,
+                                n = nrow(data) * ncol(data))
 
     ## create a cell style format for specific types at the end of the existing
     # styles. gets the reference an passes it on.
@@ -491,7 +519,7 @@ write_data2 <-function(wb, sheet, data, name = NULL,
     }
 
     sel <- which(dc == openxlsx2_celltype[["character"]]) # character
-    for (i in sel) {
+    if (length(sel)) {
       data[sel][is.na(data[sel])] <- "_openxlsx_NA"
     }
 
@@ -509,13 +537,36 @@ write_data2 <-function(wb, sheet, data, name = NULL,
     # values, but contains a string. To avoid issues, set it to the missing
     # value expression
 
-    ## fix missing characters (Otherwise NA cannot be differentiated from "NA")
-    sel <- cc$is == "<is><t>_openxlsx_NA</t></is>"
-    cc$v[sel] <- "NA"
-    cc$is[sel] <- "_openxlsx_NA"
+    ## replace NA, NaN, and Inf
+    is_na <- which(cc$is == "<is><t>_openxlsx_NA</t></is>" | cc$v == "NA")
+    if (length(is_na)) {
+      if (missing(na.strings)) {
+        cc[is_na, "v"]   <- "#N/A"
+        cc[is_na, "c_t"] <- "e"
+        cc[is_na, "is"]  <- ""
+      } else {
+        cc[is_na, "v"]  <- ""
+        if (is.null(na.strings)) {
+          # do nothing
+        } else {
+          cc[is_na, "c_t"] <- "inlineStr"
+          cc[is_na, "is"] <- txt_to_is(as.character(na.strings),
+                                        no_escapes = TRUE, raw = TRUE)
+        }
+      }
+    }
 
-    cc$v[cc$v == "NA"] <- "#N/A"
-    cc$c_t[cc$v == "#N/A"] <- "e"
+    is_nan <- which(cc$v == "NaN")
+    if (length(is_nan)) {
+      cc[is_nan, "v"]   <- "#VALUE!"
+      cc[is_nan, "c_t"] <- "e"
+    }
+
+    is_inf <- which(cc$v == "-Inf" | cc$v == "Inf")
+    if (length(is_inf)) {
+      cc[is_inf, "v"]   <- "#NUM!"
+      cc[is_inf, "c_t"] <- "e"
+    }
 
     cc$c_s[cc$typ == "0"]  <- wb$styles_mgr$get_xf_id(short_date_fmtid)
     cc$c_s[cc$typ == "1"]  <- wb$styles_mgr$get_xf_id(long_date_fmtid)
@@ -539,7 +590,8 @@ write_data2 <-function(wb, sheet, data, name = NULL,
       dims,
       dc,
       colNames,
-      removeCellStyle
+      removeCellStyle,
+      na.strings
     )
   }
 
@@ -572,7 +624,8 @@ write_data2 <-function(wb, sheet, data, name = NULL,
 #' @param bandedCols logical. If TRUE, the columns are colour banded
 #' @param bandedCols logical. If TRUE, a data table is created
 #' @param name If not NULL, a named region is defined.
-#' @param removeCellStyle keep the cell style?
+#' @param removeCellStyle if writing into existing cells, should the cell style be removed?
+#' @param na.strings optional na.strings argument. if missing #N/A is used. If NULL no cell value is written, if character or numeric this is written (even if NA is part of numeric data)
 #' @noRd
 write_data_table <- function(
     wb,
@@ -594,8 +647,9 @@ write_data_table <- function(
     bandedRows = TRUE,
     bandedCols = FALSE,
     name = NULL,
-    removeCellStyle = TRUE,
-    data_table = FALSE
+    removeCellStyle = FALSE,
+    data_table = FALSE,
+    na.strings
 ) {
 
   op <- openxlsx2_options()
@@ -618,6 +672,7 @@ write_data_table <- function(
     startRow <- min(dims[[2]])
   }
 
+  if (missing(na.strings)) na.strings <- substitute()
 
   ## common part ---------------------------------------------------------------
   if ((!is.character(sep)) || (length(sep) != 1))
@@ -751,7 +806,8 @@ write_data_table <- function(
     rowNames = rowNames,
     startRow = startRow,
     startCol = startCol,
-    removeCellStyle = removeCellStyle
+    removeCellStyle = removeCellStyle,
+    na.strings = na.strings
   )
 
 
@@ -836,7 +892,8 @@ write_data_table <- function(
 #' @param withFilter If `TRUE`, add filters to the column name row. NOTE can only have one filter per worksheet.
 #' @param sep Only applies to list columns. The separator used to collapse list columns to a character vector e.g. sapply(x$list_column, paste, collapse = sep).
 #' @param name If not NULL, a named region is defined.
-#' @param removeCellStyle keep the cell style?
+#' @param removeCellStyle if writing into existing cells, should the cell style be removed?
+#' @param na.strings optional na.strings argument. if missing #N/A is used. If NULL no cell value is written, if character or numeric this is written (even if NA is part of numeric data)
 #' @seealso [write_datatable()]
 #' @export write_data
 #' @details Formulae written using write_formula to a Workbook object will not get picked up by read_xlsx().
@@ -918,8 +975,12 @@ write_data <- function(
     withFilter = FALSE,
     sep = ", ",
     name = NULL,
-    removeCellStyle = TRUE
+    removeCellStyle = FALSE,
+    na.strings
 ) {
+
+  if (missing(na.strings)) na.strings <- substitute()
+
   write_data_table(
     wb = wb,
     sheet = sheet,
@@ -941,7 +1002,8 @@ write_data <- function(
     bandedCols = FALSE,
     name = name,
     removeCellStyle = removeCellStyle,
-    data_table = FALSE
+    data_table = FALSE,
+    na.strings = na.strings
   )
 }
 
@@ -1103,6 +1165,7 @@ write_formula <- function(wb,
 #' @param lastColumn logical. If TRUE, the last column is bold
 #' @param bandedRows logical. If TRUE, rows are colour banded
 #' @param bandedCols logical. If TRUE, the columns are colour banded
+#' @param na.strings optional na.strings argument. if missing #N/A is used. If NULL no cell value is written, if character or numeric this is written (even if NA is part of numeric data)
 #' @details columns of x with class Date/POSIXt, currency, accounting,
 #' hyperlink, percentage are automatically styled as dates, currency, accounting,
 #' hyperlinks, percentages respectively.
@@ -1225,8 +1288,12 @@ write_datatable <- function(
     firstColumn = FALSE,
     lastColumn = FALSE,
     bandedRows = TRUE,
-    bandedCols = FALSE
+    bandedCols = FALSE,
+    na.strings
 ) {
+
+  if (missing(na.strings)) na.strings <- substitute()
+
   write_data_table(
     wb = wb,
     sheet = sheet,
@@ -1248,6 +1315,7 @@ write_datatable <- function(
     bandedCols = bandedCols,
     name = NULL,
     removeCellStyle = FALSE,
-    data_table = TRUE
+    data_table = TRUE,
+    na.strings = na.strings
   )
 }